add velocity

+++ src/Particle.hpp
@@ -4,7 +4,7 @@
 class Particle {
 public:
     ParticleType pType; // type of particle
-    sf::Vector2i velocity;
+    sf::Vector2f velocity;
     uint8_t lifeTime; // time particle will live in frames (only used in fire)
     bool hasBeenUpdated; // true if particle has been updated this frame
